--- .prebuildrc
@@ -1,6 +1,0 @@
-{
-  "runtime": "napi",
-  "include-regex": "(sharp-.+\\.node|libvips-.+\\.dll)",
-  "prerelease": true,
-  "strip": true
-}
